+++ Cogs/Peep.py
@@ -165,20 +165,20 @@ class Peep(commands.Cog):
             return
         lib.sql.add_member(interaction.user)
         con = sqlite3.connect(lib.get.database_path())
-        sender_peeps = con.execute("""
-        SELECT (caught_peeps)
+        sender_total_peeps, sender_sent_peeps = con.execute("""
+        SELECT (caught_peeps, sent_peeps)
         FROM members
         WHERE guild_id = ?
         AND user_id = ?
         """, (interaction.guild_id, interaction.user.id)).fetchone()[0]
-        if amount > sender_peeps:
+        if amount > sender_total_peeps:
             con.close()
             await interaction.response.send_message("You don't have that many Peeps to transfer")
-            logging.peep_transfer("Member tried to transfer more Peeps than they have", interaction, amount, recipient.id, sender_peeps)
+            logging.peep_transfer("Member tried to transfer more Peeps than they have", interaction, amount, recipient.id, sender_total_peeps)
             return
         lib.sql.add_member(recipient)
-        receiver_peeps = con.execute("""
-        SELECT (caught_peeps)
+        receiver_total_peeps, receiver_received_peeps = con.execute("""
+        SELECT (caught_peeps, received_peeps)
         FROM members
         WHERE guild_id = ?
         AND user_id = ?
@@ -192,7 +192,7 @@ class Peep(commands.Cog):
             received_peeps = ?
         WHERE guild_id = ?
         AND user_id = ?
-        """, ((receiver_peeps + amount), amount, interaction.guild_id, recipient.id))
+        """, ((receiver_total_peeps + amount), (receiver_received_peeps + amount), interaction.guild_id, recipient.id))
         # update sending Member
         con.execute("""
         UPDATE members
@@ -201,11 +201,11 @@ class Peep(commands.Cog):
             sent_peeps = ?
         WHERE guild_id = ?
         AND user_id = ?;
-        """, ((sender_peeps - amount), amount, interaction.guild_id, interaction.user.id))
+        """, ((sender_total_peeps - amount), (sender_sent_peeps + amount), interaction.guild_id, interaction.user.id))
         con.commit()
         con.close()
         await interaction.response.send_message(f"You transferred {amount} Peeps to <@{recipient.id}>")
-        logging.peep_transfer("Member transferred Peeps to another Member", interaction, amount, recipient.id, sender_peeps, receiver_peeps)
+        logging.peep_transfer("Member transferred Peeps to another Member", interaction, amount, recipient.id, sender_total_peeps, receiver_total_peeps)
 
 
 async def setup(bot) -> None: